requirements.txt
Django>=1.6,<1.8env-excavator==1.2.0
python-dotenv==0.1.0
dj-database-url==0.3.0
djangorestframework==3.0.4
django-argonauts==1.0.1
